DELETE methods now include auth

+++ src/services/meetingService.js
@@ -87,14 +87,10 @@ module.exports = {
         }
     },
 
-    deleteMeeting: async function (meetingId) {
+    deleteMeeting: async function (userToken, meetingId) {
 
         try {
             var nowDate = new Date(Date.now());
-    
-            // TODO: Comprobar que existe un usuario autenticado y
-            // obtener su ID.
-            let fakeUserId = "1";
     
             // Obtiene de la base de datos el meeting que se
             // pretende eliminar.
@@ -105,12 +101,11 @@ module.exports = {
             
             // Comprobar que existe un meeting con la ID proporcionada.
             if (!doc) {
-                throw "Error 404: We couldn't find any meeting with the given ID."
-    
-            // Comprobar que la ID del usuario que está intentando
-            // abandonar el meeting coincide con su creatorId.
-            } else if (doc.creatorId.toString().localeCompare(fakeUserId) != 0) {
-                throw "Error 400: You can't delete a meeting that you didn't create.";
+                throw "Error 404: We couldn't find any meeting with the given ID.";
+
+            // Comprobar que existe un usuario autenticado
+            } else if (userToken == null) {
+                throw "Error 401: You must be authenticated to delete a meeting."; 
     
             // Comprobar que la startingDate del meeting es futura
             // y, por tanto, dicho meeting no ha comenzado.
@@ -118,22 +113,39 @@ module.exports = {
                 throw "Error 400: You can't delete the meeting. It has already started.";
             
             } else {
-                // El listado de miembros del meeting recoge las dependencias
-                // con los usuarios que se deben actualizar en la base de datos
-                // del microservicio profile.
-                var members = doc.members;
-    
-                // Eliminamos el meeting de nuestra base de datos.
-                await Meeting.deleteOne({_id : meetingId}).exec();
+
+                // Obtener la información del usuario, incluyendo la ID,
+                // a partir de su token.
+                let userInfo = await authAxios.get('auth/me', {headers: {'x-access-token': userToken}});
+                console.log(userInfo);
                 
-                // Actualizamos las dependencias del microservicio profile,
-                // quitando el meeting eliminado para cada uno de los miembros
-                // en sus respectivos listados de meetings. 
-                for (var i = 0; i < members.length; i++) {
-                    await profileAxios.put('user/' + members[i] + '/leavesMeeting/' + meetingId);
+                // Comprobar que existe un usuario con el token recibido.
+                if(!userInfo) {
+                    throw "Error 401: No user matches the given token.";
+                
+                // Comprobar que la ID del usuario que está intentando
+                // abandonar el meeting coincide con su creatorId.
+                } else if (doc.creatorId.toString().localeCompare(userInfo.data._id) != 0) {
+                    throw "Error 403: You can't delete a meeting that you didn't create.";
+
+                } else {
+                    // El listado de miembros del meeting recoge las dependencias
+                    // con los usuarios que se deben actualizar en la base de datos
+                    // del microservicio profile.
+                    var members = doc.members;
+        
+                    // Eliminamos el meeting de nuestra base de datos.
+                    await Meeting.deleteOne({_id : meetingId}).exec();
+                    
+                    // Actualizamos las dependencias del microservicio profile,
+                    // quitando el meeting eliminado para cada uno de los miembros
+                    // en sus respectivos listados de meetings. 
+                    for (var i = 0; i < members.length; i++) {
+                        await profileAxios.put('user/' + members[i] + '/leavesMeeting/' + meetingId);
+                    }
+        
+                    return [false, null];
                 }
-    
-                return [false, null];
             }
         
         } catch (error) {
@@ -141,14 +153,10 @@ module.exports = {
         }
     },
 
-    leaveMeeting: async function (meetingId) {
+    leaveMeeting: async function (userToken,meetingId) {
 
         try {
             var nowDate = new Date(Date.now());
-    
-            // TODO: Comprobar que existe un usuario autenticado y
-            // obtener su ID.
-            let fakeUserId = "1";
     
             // Obtiene de la base de datos el meeting que se
             // pretende abandonar.
@@ -160,38 +168,56 @@ module.exports = {
             if (!doc) {
                 throw "Error 404: We couldn't find any meeting with the given ID."
     
-            // Comprobar que la ID del usuario que está intentando
-            // abandonar el meeting no coincide con su creatorId.
-            } else if (doc.creatorId.toString().localeCompare(fakeUserId) == 0) {
-                throw "Error 400: You can't leave a meeting that you created.";
-    
             // Comprobar que la startingDate del meeting es futura
             // y, por tanto, dicho meeting no ha comenzado.
             } else if (doc.startingDate <= nowDate) {
                 throw "Error 400: You can't leave the meeting. It has already started.";
-    
-            // Comprobar que la ID del usuario autenticado se
-            // encuentra en el listado de members del meeting.
-            } else if (!doc.members.includes(fakeUserId)) {
-                throw "Error 400: You can't leave a meeting that you are not a member of.";
+
+            // Comprobar que existe un usuario autenticado
+            } else if (userToken == null) {
+                throw "Error 401: You must be authenticated to leave a meeting.";
             
             } else {
-                // Eliminar la ID del usuario autenticado del listado
-                // de members del meeting.
-                var newMembers = doc.members;
-                newMembers.splice(newMembers.indexOf(fakeUserId), 1);
-    
-                // Actualizar el meeting con su nuevo listado de
-                // members en la base de datos.
-                let updatedMeeting = await Meeting.findByIdAndUpdate(meetingId, {members: newMembers}, {new: true});
-                console.log(updatedMeeting);
-    
-                // Llamada al microservicio de Profile para actualizar
-                // las dependencias y reflejar el abandono del usuario
-                // del meeting.
-                await profileAxios.put('user/' + fakeUserId + '/leavesMeeting/' + meetingId);
-    
-                return [false, updatedMeeting];
+                // Obtener la información del usuario, incluyendo la ID,
+                // a partir de su token.
+                let userInfo = await authAxios.get('auth/me', {headers: {'x-access-token': userToken}});
+                console.log(userInfo);
+                
+                // Comprobar que existe un usuario con el token recibido.
+                if(!userInfo) {
+                    throw "Error 401: No user matches the given token.";
+                
+                // Comprobar que la ID del usuario que está intentando
+                // abandonar el meeting no coincide con su creatorId.
+                } else if (doc.creatorId.toString().localeCompare(userInfo.data._id) == 0) {
+                    throw "Error 403: You can't leave a meeting that you created.";
+
+                // Comprobar que la ID del usuario autenticado se
+                // encuentra en el listado de members del meeting.
+                } else if (!doc.members.includes(userInfo.data._id)) {
+                    throw "Error 400: You can't leave a meeting that you are not a member of.";
+
+                } else {
+                    // Obtener la ID del usuario autenticado.
+                    var userId = userInfo.data._id;
+
+                    // Eliminar la ID del usuario autenticado del listado
+                    // de members del meeting.
+                    var newMembers = doc.members;
+                    newMembers.splice(newMembers.indexOf(userId), 1);
+        
+                    // Actualizar el meeting con su nuevo listado de
+                    // members en la base de datos.
+                    let updatedMeeting = await Meeting.findByIdAndUpdate(meetingId, {members: newMembers}, {new: true});
+                    console.log(updatedMeeting);
+        
+                    // Llamada al microservicio de Profile para actualizar
+                    // las dependencias y reflejar el abandono del usuario
+                    // del meeting.
+                    await profileAxios.put('user/' + userId + '/leavesMeeting/' + meetingId);
+        
+                    return [false, updatedMeeting];
+                }
             }
         
         } catch (error) {
@@ -310,7 +336,7 @@ module.exports = {
                     // Comprobar que la ID del usuario que está intentando
                     // actualizar el meeting coincide con su creatorId.
                     } else if (doc.creatorId.toString().localeCompare(userId.toString()) != 0) {
-                        throw "Error 400: You can't update a meeting that you didn't create.";
+                        throw "Error 403: You can't update a meeting that you didn't create.";
 
                     // Comprobar que la capacidad del meeting no es menor al número de miembros
                     // ya registrados en el meeting.
